removed the need for chat context
